rename files to match class name (#37)
