+++ src/app/modules/core/interceptors/jwt.interceptor.spec.ts
@@ -3,7 +3,7 @@ import { HttpClientTestingModule, HttpTestingController } from '@angular/common/
 import { HTTP_INTERCEPTORS, HttpClient } from '@angular/common/http';
 
 import { JwtInterceptor } from './jwt.interceptor';
-import { AuthenticationService } from '../services/auth.service';
+import { AuthenticationService } from '../services/authentication.service';
 import { RouterTestingModule } from '@angular/router/testing';
 
 describe('JwtInterceptor', () => {
