+++ src/app/modules/onboarding/pages/hd-wallet-wizard/hd-wallet-wizard.component.spec.ts
@@ -12,7 +12,7 @@ import { GenerateMnemonicComponent } from '../../components/generate-mnemonic/ge
 import { ConfirmMnemonicComponent } from '../../components/confirm-mnemonic/confirm-mnemonic.component';
 import { GenerateAccountsComponent } from '../../components/generate-accounts/generate-accounts.component';
 import { WalletService } from 'src/app/modules/core/services/wallet.service';
-import { AuthenticationService } from 'src/app/modules/core/services/auth.service';
+import { AuthenticationService } from 'src/app/modules/core/services/authentication.service';
 import { WalletResponse, AuthResponse, CreateWalletRequest } from 'src/app/proto/validator/accounts/v2/web_api';
 
 describe('HdWalletWizardComponent', () => {
